% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_observed_data_loglik.R
\name{cv_observed_data_loglik}
\alias{cv_observed_data_loglik}
\title{Cross-validated observed-data log-likelihood
This function returns the value of the observed-data log-likelihood based on cross-validation.}
\usage{
cv_observed_data_loglik(
  q = 4,
  sN = 10,
  fold,
  Y_unval = NULL,
  Y_val = NULL,
  X_unval = NULL,
  X_val = NULL,
  C = NULL,
  Validated = NULL,
  Bspline = NULL,
  data,
  theta_pred = NULL,
  gamma_pred = NULL,
  TOL = 1e-04,
  MAX_ITER = 1000
)
}
\arguments{
\item{q}{Degree of the B-spline basis. Default is \code{q = 4} for a cubic B-spline basis.}

\item{sN}{Total number of B-spline basis functions. Defualt is \code{sN = 10} functions.}

\item{fold}{Column name with the assigned fold for cross-validation.}

\item{Y_unval}{Column name with the unvalidated outcome. If \code{Y_unval} is null, the outcome is assumed to be error-free.}

\item{Y_val}{Column name with the validated outcome.}

\item{X_unval}{Column name(s) with the unvalidated predictors.  If \code{X_unval} and \code{X_val} are \code{null}, all precictors are assumed to be error-free.}

\item{X_val}{Column name(s) with the validated predictors. If \code{X_unval} and \code{X_val} are \code{null}, all precictors are assumed to be error-free.}

\item{C}{(Optional) Column name(s) with additional error-free covariates.}

\item{Validated}{Column name with the validation indicator. The validation indicator can be defined as \code{Validated = 1} or \code{TRUE} if the subject was validated and \code{Validated = 0} or \code{FALSE} otherwise.}

\item{Bspline}{Vector of column names containing the B-spline basis functions.}

\item{data}{A dataframe with one row per subject containing columns: \code{Y_unval}, \code{Y_val}, \code{X_unval}, \code{X_val}, \code{C}, \code{Validated}, and \code{Bspline}.}

\item{theta_pred}{Vector of columns in \code{data} that pertain to the predictors in the analysis model.}

\item{gamma_pred}{Vector of columns in \code{data} that pertain to the predictors in the outcome error model.}

\item{TOL}{Tolerance between iterations in the EM algorithm used to define convergence.}

\item{MAX_ITER}{Maximum number of iterations allowed in the EM algorithm.}
}
\value{
scalar value of the function
}
\description{
Cross-validated observed-data log-likelihood
This function returns the value of the observed-data log-likelihood based on cross-validation.
}
