% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_theta.R
\name{pl_theta}
\alias{pl_theta}
\title{Profile likelihood for theta, the analysis model parameters}
\usage{
pl_theta(
  k,
  theta,
  h_N,
  n,
  N,
  Y_unval,
  Y_val,
  X_unval,
  X_val,
  C,
  Bspline,
  comp_dat_all,
  theta_pred,
  gamma_pred,
  gamma0 = NULL,
  p0 = NULL,
  p_val_num = NULL,
  TOL,
  MAX_ITER
)
}
\arguments{
\item{k}{A numeric index between 1 and the dimension of theta for the element of theta to be perturbed}

\item{theta}{Parameters for the analysis model (a column vector) at convergence, resulting from the EM algorithm}

\item{h_N}{Size of the small perturbation in \code{theta[k]}, by default chosen to be \code{h_N =  N ^ ( - 1 / 2)}}

\item{n}{Phase II sample size}

\item{N}{Phase I sample size}

\item{Y_unval}{Column names with the unvalidated outcome. If \code{Y_unval} is null, the outcome is assumed to be error-free.}

\item{Y_val}{Column names with the validated outcome.}

\item{X_unval}{Column name(s) with the unvalidated predictors.  If \code{X_unval} and \code{X_val} are \code{null}, all precictors are assumed to be error-free.}

\item{X_val}{Column name(s) with the validated predictors. If \code{X_unval} and \code{X_val} are \code{null}, all precictors are assumed to be error-free.}

\item{C}{(Optional) Column name(s) with additional error-free covariates.}

\item{Bspline}{Vector of column names containing the B-spline basis functions.}

\item{comp_dat_all}{Augmented dataset containing rows for each combination of unvalidated subjects' data with values from Phase II (a matrix)}

\item{theta_pred}{Vector of columns in \code{comp_dat_all} that pertain to the predictors in the analysis model.}

\item{gamma_pred}{Vector of columns in \code{comp_dat_all} that pertain to the predictors in the outcome error model.}

\item{gamma0}{Starting values for \code{gamma}, the parameters for the outcome error model (a column vector)}

\item{p0}{Starting values for \code{p}, the B-spline coefficients for the approximated covariate error model (a matrix)}

\item{p_val_num}{Contributions of validated subjects to the numerator for \code{p}, which are fixed (a matrix)}

\item{TOL}{Tolerance between iterations in the EM algorithm used to define convergence.}

\item{MAX_ITER}{Maximum number of iterations allowed in the EM algorithm.}

\item{Validated}{Column name with the validation indicator. The validation indicator can be defined as \code{Validated = 1} or \code{TRUE} if the subject was validated and \code{Validated = 0} or \code{FALSE} otherwise.}
}
\value{
Profile likelihood for \code{theta} after perturbing element \code{k} by \code{h_N}.
}
\description{
This function returns the value of the profile log-likelihood for parameters \code{theta} of the
analysis model P(Y|X,C) after perturbing element \code{k} of \code{theta} by some small amount \code{h_N}.
}
