% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_out_rw.R
\name{profile_out_rw}
\alias{profile_out_rw}
\title{Profiles out nuisance parameters from the observed-data log-likelihood for a given value of theta}
\usage{
profile_out_rw(
  theta,
  n,
  N,
  Y_unval = NULL,
  Y_val = NULL,
  X_unval = NULL,
  X_val = NULL,
  C = NULL,
  Bspline = NULL,
  comp_dat_all,
  theta_pred,
  gamma_pred,
  gamma0,
  p0,
  p_val_num,
  TOL,
  MAX_ITER
)
}
\arguments{
\item{theta}{Parameters for the analysis model (a column vector)}

\item{n}{Phase II sample size}

\item{N}{Phase I sample size}

\item{Y_unval}{Column with the unvalidated outcome (can be name or numeric index)}

\item{Y_val}{Column with the validated outcome (can be name or numeric index)}

\item{X_unval}{Column(s) with the unvalidated predictors (can be name or numeric index)}

\item{X_val}{Column(s) with the validated predictors (can be name or numeric index)}

\item{C}{(Optional) Column(s) with additional error-free covariates (can be name or numeric index)}

\item{Bspline}{Vector of columns containing the B-spline basis functions (can be name or numeric index)}

\item{comp_dat_all}{Augmented dataset containing rows for each combination of unvalidated subjects' data with values from Phase II (a matrix)}

\item{theta_pred}{Vector of columns in \code{data} that pertain to the predictors in the analysis model.}

\item{gamma_pred}{Vector of columns in \code{data} that pertain to the predictors in the outcome error model.}

\item{gamma0}{Starting values for \code{gamma}, the parameters for the outcome error model (a column vector)}

\item{p0}{Starting values for \code{p}, the B-spline coefficients for the approximated covariate error model (a matrix)}

\item{p_val_num}{Contributions of validated subjects to the numerator for \code{p}, which are fixed (a matrix)}

\item{TOL}{Tolerance between iterations in the EM algorithm used to define convergence.}

\item{MAX_ITER}{Maximum number of iterations allowed in the EM algorithm.}
}
\value{
Profile likelihood for \code{theta}: the value of the observed-data log-likelihood after profiling out other parameters.
}
\description{
For a given vector \code{theta} to parameterize P(Y|X,C), this function repeats the EM algorithm to find
the values of \code{gamma} and \code{p} at convergence. The resulting parameters are used to find the profile
log-likelihood for \code{theta} by plugging them into the observed-data log-likelihood.
This function is used by \code{pl_theta()}.
}
