% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg2ph.R
\name{logreg2ph}
\alias{logreg2ph}
\title{Sieve maximum likelihood estimator (SMLE) for two-phase logistic regression problems}
\usage{
logreg2ph(
  Y_unval = NULL,
  Y_val = NULL,
  X_unval = NULL,
  X_val = NULL,
  C = NULL,
  Validated = NULL,
  Bspline = NULL,
  data,
  initial_lr_params = "Zero",
  h_n_scale = NULL,
  noSE = FALSE,
  TOL = 1e-04,
  MAX_ITER = 1000
)
}
\arguments{
\item{Y_unval}{Column with the unvalidated outcome (can be name or numeric index)}

\item{Y_val}{Column with the validated outcome (can be name or numeric index)}

\item{X_unval}{Column(s) with the unvalidated predictors (can be name or numeric index)}

\item{X_val}{Column(s) with the validated predictors (can be name or numeric index)}

\item{C}{(Optional) Column(s) with additional error-free covariates (can be name or numeric index)}

\item{Validated}{Columns with the validation indicator (can be name or numeric index)}

\item{Bspline}{Vector of columns containing the B-spline basis functions (can be name or numeric index)}

\item{data}{A dataframe with one row per subject containing columns: Y_unval, Y_val, X_unval, X_val, C, Validated, and Bspline.}

\item{initial_lr_params}{Initial values for parametric model parameters. Choices include (1) "Zero" (non-informative starting values) or (2) "Complete-data" (estimated based on validated subjects only)}

\item{h_n_scale}{Size of the perturbation used in estimating the standard errors via profile likelihood. If none is supplied, defaults to 1.}

\item{noSE}{Indicator for whether standard errors are desired. Defaults to noSE = FALSE.}

\item{TOL}{Tolerance between iterations in the EM algorithm used to define convergence.}

\item{MAX_ITER}{Maximum number of iterations allowed in the EM algorithm.}
}
\value{
List containing (1) Coefficients: a dataframe with final coefficient and standard error estimates (where applicable), (2) converged: indicator of estimator convergence, (3) se_converged: indicator of standard error estimator convergence, (4) converged_msg: (where applicable) description of non-convergence, (5) iterations: number of iterations completed by EM algorithm
}
\description{
This function returns the value of the observed-data log-likelihood (equation (2) in Lotspeich et al. (2020))
for a given dataset and parameter values \code{theta}, \code{gamma}, and \code{p}.
}
