% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_data_loglik.R
\name{observed_data_loglik}
\alias{observed_data_loglik}
\title{Observed-data log-likelihood}
\usage{
observed_data_loglik(
  N,
  n,
  Y_unval = NULL,
  Y_val = NULL,
  X_unval = NULL,
  X_val = NULL,
  C = NULL,
  Bspline = NULL,
  comp_dat_all,
  theta_pred,
  gamma_pred,
  theta,
  gamma,
  p
)
}
\arguments{
\item{N}{Phase I sample size}

\item{n}{Phase II sample size}

\item{Y_unval}{Column with the unvalidated outcome (can be name or numeric index)}

\item{Y_val}{Column with the validated outcome (can be name or numeric index)}

\item{X_unval}{Column(s) with the unvalidated predictors (can be name or numeric index)}

\item{X_val}{Column(s) with the validated predictors (can be name or numeric index)}

\item{C}{(Optional) Column(s) with additional error-free covariates (can be name or numeric index)}

\item{Bspline}{Vector of columns containing the B-spline basis functions (can be name or numeric index)}

\item{comp_dat_all}{Augmented dataset containing rows for each combination of unvalidated subjects' data with values from Phase II (a matrix)}

\item{theta_pred}{Vector of columns in \code{data} that pertain to the predictors in the analysis model.}

\item{gamma_pred}{Vector of columns in \code{data} that pertain to the predictors in the outcome error model.}

\item{theta}{Parameters for the analysis model (a column vector)}

\item{gamma}{Parameters for the outcome error model (a column vector)}

\item{p}{B-spline coefficients for the approximated covariate error model (a matrix)}
}
\value{
Scalar value of the function
}
\description{
This function returns the value of the observed-data log-likelihood (equation (2) in Lotspeich et al. (2020))
for a given dataset and parameter values \code{theta}, \code{gamma}, and \code{p}.
}
